refactor: 명확한 bust 기준 네이밍으로 수정 및 불필요한 호출 제거
