+++ src/main/kotlin/blackjack/event/GameEvent.kt
@@ -1,7 +1,9 @@
-package blackjack.domain.game
+package blackjack.event
+
+import blackjack.domain.player.Player
 
 private typealias HitOrNot = (String) -> Boolean
-private typealias ResultEvent = (String, List<Pair<String, String>>) -> Unit
+private typealias ResultEvent = (Player) -> Unit
 
 class GameEvent(
     val hitOrNot: HitOrNot,
